refactor: extract shared store module and add go workspace
Move the duplicated store package (Deployment types, Status constants,
JSONStore persistence engine) into a new shared module at store/ with
path github.com/ercadev/dirigent/store.

Both api and orchestrator now import from the shared module instead of
maintaining their own copies. A go.work workspace wires all three
modules together for local development. Explicit replace directives are
also added to each go.mod so builds work correctly with Go 1.25's
stricter proxy verification for workspace modules.

Also updates README.md and api/README.md to reflect the current
monorepo structure (api/, orchestrator/, store/, dashboard/).
