+++ api/go.mod
@@ -1,3 +1,7 @@
 module github.com/ercadev/dirigent
 
 go 1.24.0
+
+require github.com/ercadev/dirigent/store v0.0.0
+
+replace github.com/ercadev/dirigent/store => ../store
