+++ go.work
@@ -1,0 +1,7 @@
+go 1.24.0
+
+use (
+	./api
+	./orchestrator
+	./store
+)
