+++ store/go.mod
@@ -1,0 +1,3 @@
+module github.com/ercadev/dirigent/store
+
+go 1.24.0
